+++ server/src/main/java/server/model/Card.java
@@ -1,8 +1,9 @@
 package server.model;
 
-import common.Auth;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
+
 import lombok.NoArgsConstructor;
 import server.auth.Role;
 import server.auth.Status;
@@ -15,9 +16,19 @@ import java.math.BigDecimal;
 @NoArgsConstructor
 public class Card implements Auth {
     private Long id;
-    private String login;
-    private String password;
+    private String cardNumber;
+    private String cardPassword;
     private BigDecimal balance;
     private Role cardRole = Role.USER;
     private Status status;
+
+    @Override
+    public String getLogin() {
+        return cardNumber;
+    }
+
+    @Override
+    public String getPassword() {
+        return cardPassword;
+    }
 }
