+++ server/src/main/java/server/model/Person.java
@@ -6,16 +6,16 @@ import lombok.NoArgsConstructor;
 import server.auth.Role;
 import server.auth.Status;
 
-import java.math.BigDecimal;
+import java.util.Set;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class Person {
+public class Person implements Auth{
     private Long id;
-    private String cardNumber;
-    private String cardPassword;
-    private BigDecimal balance;
+    private String login;
+    private String password;
+    private Set<Card> cards;
     private Role cardRole;
     private Status status;
 }
